UBUNTU: SAUCE: signaltest: fix a build warnings on loff_t print specifier
The loff_t is warning on both 32 and 64 bit builds with newer versions
of gcc, so fix this by casting it up to a long long.

+++ rmaptest/src/rmap-test.c
@@ -194,7 +194,7 @@ int main(int argc, char *argv[])
 	file_size *= pagesize;
 
 	printf("Total file size: %lldk, Total memory: %lldk\n",
-		file_size / 1024,
+		(long long)file_size / 1024,
 		((long long)nvmas * vmasize * pagesize) / 1024);
 
 	if (ftruncate(fd, file_size) < 0) {
